megatron-moe-model/requirements.txt
torch>=2.1.0
numpy>=1.24.0
pyyaml>=6.0
triton>=2.1.0; platform_system == "Linux"
